moved greeting
