+++ app/src/main/java/com/example/jetpacklist/model/LandmarkModelData.kt
@@ -1,0 +1,5 @@
+package com.example.jetpacklist.model
+
+data class LandmarkModelData(
+    val id:Int
+)
