+++ app/src/main/java/com/example/jetpacklist/views/LandmarkDetail.kt
@@ -1,0 +1,4 @@
+package com.example.jetpacklist.views
+
+class LandmarkDetail {
+}
